+++ hive/replays/__init__.py
@@ -1,6 +1,6 @@
 from hive.replays.circular_replay import CircularReplayBuffer, SimpleReplayBuffer
 from hive.replays.legal_moves_replay import LegalMovesBuffer
-from hive.replays.ppo_replay import OnPolicyReplayBuffer
+from hive.replays.on_policy_replay import OnPolicyReplayBuffer
 from hive.replays.prioritized_replay import PrioritizedReplayBuffer
 from hive.replays.recurrent_replay import RecurrentReplayBuffer
 from hive.replays.replay_buffer import BaseReplayBuffer
